+++ monglife-module-mqtt/src/main/java/com/monglife/module/mqtt/annotation/MqttMapping.java
