+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'monglife-module-mqtt'
+
+include 'monglife-module-mqtt-api'
+include 'monglife-module-mqtt'
