--- src/main/java/com/monglife/module/mqtt/client/MqttOutBoundClient.java
@@ -1,13 +1,0 @@
-package com.monglife.module.mqtt.client;
-
-import org.springframework.integration.annotation.MessagingGateway;
-import org.springframework.integration.mqtt.support.MqttHeaders;
-import org.springframework.messaging.handler.annotation.Header;
-import org.springframework.stereotype.Component;
-
-@Component
-@MessagingGateway(defaultRequestChannel = "mqttOutboundChannel")
-public interface MqttOutBoundClient {
-
-    void send(@Header(MqttHeaders.TOPIC) String topic, String payload);
-}
